Replaced busfinder with backbone implementation
