--- web/templates/busfinder-backbone.html
@@ -1,110 +1,0 @@
-<html>
-<head>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="/static/css/jquery.mobile-1.3.0-rc.1.min.css" />
-	<style type="text/css">
-		.ui-content {
-			padding: 5px 15px 0px 15px;
-		}
-		#stopSearchResults {
-			text-align: center;
-		}
-		#stopSearchResults #stopName {
-			font-size: 18px;
-			font-weight: bold;
-		}
-	</style>
-</head>
-<body>
-    <div id="header" data-role="header">
-        <h3>HRT Bus Finder</h3>
-    </div>
-
-    <div id="content" data-role="content"></div>
-    <div id="mapcanvas" style="height: 100%; width: 100%;"></div>
-
-	<script src="/static/js/jquery-1.9.1.js"></script>
-	<script type="text/javascript">
-		$(document).on("mobileinit", function () {
-		    console.log('mobileinit')
-		    $.mobile.ajaxEnabled = false;
-		    // Let Backbone handle click events 
-		    $.mobile.linkBindingEnabled = false;
-		    $.mobile.listview.prototype.options.icon = "";
-		    // Don't listen to hash changes. Let backbone do this
-		    $.mobile.hashListeningEnabled = false;
-		    $.mobile.pushStateEnabled = false;
-		    // No hoverdelay for buttons. Use FastClick for 0 delay a tags
-		    $.mobile.buttonMarkup.hoverDelay = 0;
-		});
-	</script>
-	
-	<script src="http://maps.google.com/maps/api/js?sensor=true"></script>
-    <script src="/static/js/jquery.mobile-1.3.0-rc.1.js"></script>
-	<script src="/static/js/underscore-min.js"></script>
-	<script src="/static/js/backbone-min.js"></script>
-	<script src="/static/js/busfinder-backbone.js"></script>
-	
-	<script type="text/template" id="home-template">
-		<button id="findStop">Find My Stop</button>
-		<button id="goToRoutes">Skip To Routes</button>
-	</script>
-	
-	<script type="text/template" id="begin-stop-search-template">
-		<button id="stopSearch-location">Use Current Location</button>
-		<button id="stopSearch-intersection">Use Intersection</button>
-	</script>
-	
-	<script type="text/template" id="stop-search-template">
-		Intersection:<input id="intersection" type="text" placeholder="High and Court"/>
-		City: <input id="city" type="text" placeholder="Portsmouth"/>
-		<button id="search">Search</button>
-	</script>
-	
-	<script type="text/template" id="stop-search-results-template">
-		<div style="text-align:center"><b><%= stopName %><% if(stopId != 0) { %> (Stop <%= stopId %>)<% } %></b></div>
-		
-		<select id="route">
-			<option selected="selected" value="">Select Route</option>
-			
-			<optgroup label="Inbound" id="1">
-			<% _.each(inboundRoutes, function(route) { %>
-				<option value="<%= route.route_id %>"><%= route.route_short_name %> - <%= route.route_long_name %></option>
-			<% }); %>
-			</optgroup>
-			
-			<optgroup label="Outbound" id="0">
-			<% _.each(outboundRoutes, function(route) { %>
-				<option value="<%= route.route_id %>"><%= route.route_short_name %> - <%= route.route_long_name %></option>
-			<% }); %>
-			</optgroup>
-		</select>
-		
-		<div class="ui-grid-b">
-			<div class="ui-block-a"><button id="prev-stop" type="button" data-theme="c" data-icon="arrow-l">Previous</button></div>
-			<div class="ui-block-b" style="text-align:center;"><p><%= curStop %> / <%= numStops %></p></div>
-			<div class="ui-block-b"><button id="next-stop" type="button" data-theme="c" data-icon="arrow-r" data-iconpos="right">Next</button></div>	   
-		</div>
-	</script>
-	
-	<script type="text/template" id="route-view-template">
-		<select id="route">
-			<option selected="selected" value="">Select Route</option>
-			<% _.each(routes, function(route) { %>
-				<option value="<%= route.route_id %>"><%= route.route_short_name %> - <%= route.route_long_name %></option>
-			<% }); %>
-		</select>
-		<div id="routeInfo">Buses Found: <%= buses %> </div>
-	</script>
-	
-	<script type="text/template" id="stop-times-template">
-		<div style="text-align:center"><b>Bus <% if(busId) { %><%= busId %> <% } %><%= stopTimeMsg %></b></div>
-		
-		<div class="ui-grid-b">
-			<div class="ui-block-a"><button id="prev" type="button" data-theme="c" data-icon="arrow-l">Previous</button></div>
-			<div class="ui-block-b" style="text-align:center;"><p><%= curItem %> / <%= numItems %></p></div>
-			<div class="ui-block-b"><button id="next" type="button" data-theme="c" data-icon="arrow-r" data-iconpos="right">Next</button></div>	   
-		</div>
-	</script>
-</body>
-</html>
