+++ web/templates/busfinder.html
@@ -1,10 +1,10 @@
-﻿<html>
+<html>
 <head>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="/static/css/jquery.mobile-1.3.0-beta.1.min.css" />
+    <link rel="stylesheet" href="/static/css/jquery.mobile-1.3.0-rc.1.min.css" />
 	<style type="text/css">
 		.ui-content {
-			padding: 0px 15px;
+			padding: 5px 15px 0px 15px;
 		}
 		#stopSearchResults {
 			text-align: center;
@@ -14,56 +14,97 @@
 			font-weight: bold;
 		}
 	</style>
-    <script src="http://maps.google.com/maps/api/js?sensor=true"></script>
-	<script src="/static/js/jquery-1.9.0.min.js"></script>
-    <script src="/static/js/jquery.mobile-1.3.0-beta.1.min.js"></script>
-	<script src="/static/js/bus.js"></script>
-	<script src="/static/js/stop.js"></script>
-	<script src="/static/js/busfinder.js"></script>
 </head>
 <body>
     <div id="header" data-role="header">
         <h3>HRT Bus Finder</h3>
     </div>
-    <div id="content" data-role="content">
-		<div id="useStopOption" class="option">
-			<button onclick="showContent('stopSearchOption')">Find My Stop</button>
-			<button onclick="showContent('routeSelect')">Skip To Routes</button>
-		</div>
+
+    <div id="content" data-role="content"></div>
+    <div id="mapcanvas" style="height: 100%; width: 100%;"></div>
+
+	<script src="/static/js/jquery-1.9.1.min.js"></script>
+	<script type="text/javascript">
+		$(document).on("mobileinit", function () {
+		    console.log('mobileinit')
+		    $.mobile.ajaxEnabled = false;
+		    // Let Backbone handle click events 
+		    $.mobile.linkBindingEnabled = false;
+		    $.mobile.listview.prototype.options.icon = "";
+		    // Don't listen to hash changes. Let backbone do this
+		    $.mobile.hashListeningEnabled = false;
+		    $.mobile.pushStateEnabled = false;
+		    // No hoverdelay for buttons. Use FastClick for 0 delay a tags
+		    $.mobile.buttonMarkup.hoverDelay = 0;
+		});
+	</script>
+	
+	<script src="http://maps.google.com/maps/api/js?sensor=true"></script>
+    <script src="/static/js/jquery.mobile-1.3.0-rc.1.min.js"></script>
+	<script src="/static/js/underscore-min.js"></script>
+	<script src="/static/js/backbone-min.js"></script>
+	<script src="/static/js/busfinder.js"></script>
+	
+	<script type="text/template" id="home-template">
+		<button id="findStop">Find My Stop</button>
+		<button id="goToRoutes">Skip To Routes</button>
+	</script>
+	
+	<script type="text/template" id="begin-stop-search-template">
+		<button id="stopSearch-location">Use Current Location</button>
+		<button id="stopSearch-intersection">Use Intersection</button>
+	</script>
+	
+	<script type="text/template" id="stop-search-template">
+		Intersection:<input id="intersection" type="text" placeholder="High and Court"/>
+		City: <input id="city" type="text" placeholder="Portsmouth"/>
+		<button id="search">Search</button>
+	</script>
+	
+	<script type="text/template" id="stop-search-results-template">
+		<div style="text-align:center"><b><%= stopName %><% if(stopId != 0) { %> (Stop <%= stopId %>)<% } %></b></div>
 		
-		<div id="stopSearchOption"  class="option" style="display:none;">
-			<button onclick="locationSearchOnPosition()">Use Current Location</button>
-			<button onclick="showContent('stopSearch')">Use Intersection</button>
-		</div>
+		<select id="route">
+			<option selected="selected" value="">Select Route</option>
+			
+			<optgroup label="Inbound" id="1">
+			<% _.each(inboundRoutes, function(route) { %>
+				<option value="<%= route.route_id %>"><%= route.route_short_name %> - <%= route.route_long_name %></option>
+			<% }); %>
+			</optgroup>
+			
+			<optgroup label="Outbound" id="0">
+			<% _.each(outboundRoutes, function(route) { %>
+				<option value="<%= route.route_id %>"><%= route.route_short_name %> - <%= route.route_long_name %></option>
+			<% }); %>
+			</optgroup>
+		</select>
 		
-		<div id="stopSearch" class="option" style="display:none;">
-			Intersection:<input id="intersection" type="text" placeholder="High and Court"/>
-			City: <input id="city" type="text" placeholder="Portsmouth"/>
-			<button onclick="locationSearchOnIntersection()">Search</button>
+		<div class="ui-grid-b">
+			<div class="ui-block-a"><button id="prev-stop" type="button" data-theme="c" data-icon="arrow-l">Previous</button></div>
+			<div class="ui-block-b" style="text-align:center;"><p><%= curStop %> / <%= numStops %></p></div>
+			<div class="ui-block-b"><button id="next-stop" type="button" data-theme="c" data-icon="arrow-r" data-iconpos="right">Next</button></div>	   
 		</div>
+	</script>
+	
+	<script type="text/template" id="route-view-template">
+		<select id="route">
+			<option selected="selected" value="">Select Route</option>
+			<% _.each(routes, function(route) { %>
+				<option value="<%= route.route_id %>"><%= route.route_short_name %> - <%= route.route_long_name %></option>
+			<% }); %>
+		</select>
+		<div id="routeInfo">Buses Found: <%= buses %> </div>
+	</script>
+	
+	<script type="text/template" id="stop-times-template">
+		<div style="text-align:center"><b>Bus <% if(busId) { %><%= busId %> <% } %><%= stopTimeMsg %></b></div>
 		
-		<div id="stopSearchResults" class="option" style="display:none;">
-			<div id="stopName">Loading...</div>
-			<div id="stopId">Select a route</div>
-			<div id="inboundRoutes">Inbound: <span class="routes"></span></div>
-			<div id="outboundRoutes">Outbound: <span class="routes"></span></div>
-			<div class="ui-grid-a">
-				<div class="ui-block-a"><button type="button" data-theme="c" data-icon="arrow-l">Previous</button></div>
-				<div class="ui-block-b"><button type="button" data-theme="c" data-icon="arrow-r" data-iconpos="right">Next</button></div>	   
-			</div>
+		<div class="ui-grid-b">
+			<div class="ui-block-a"><button id="prev" type="button" data-theme="c" data-icon="arrow-l">Previous</button></div>
+			<div class="ui-block-b" style="text-align:center;"><p><%= curItem %> / <%= numItems %></p></div>
+			<div class="ui-block-b"><button id="next" type="button" data-theme="c" data-icon="arrow-r" data-iconpos="right">Next</button></div>	   
 		</div>
-		
-		<div id="routeSelect" class="option" style="display:none;">
-			<select id="routeNumber" onchange="showBuses()">
-				<option selected="selected" value="">Select Route</option>
-			</select>
-			<div id="routeInfo">Buses Found: <span class="numBuses"></span></div>
-			<div id="scheduledStopInfo" style="display:none;">Stop <span class="stopId"></span>: <span class="details"></span></div>
-		</div>
-		
-		<div id="msg"></div>
-    </div>
-    
-    <div id="mapcanvas" style="height: 400px; width: 100%;"></div>
+	</script>
 </body>
 </html>
